\name{dat3m}
\alias{dat3m}
\docType{data}
\title{
Serialized Dose-Response D optimal experimental design model file 
}
\description{
A character string for a serialized DR design, i.e. all results from optDesign() function of the DoseFinding package given the parameters set in data(dat3i).
}
\usage{data("dat3m")}
\format{A data string
}
\details{Example experimental design model produced by exp.design.dr function
}
\source{
}
\references{https://cran.r-project.org/web/packages/DoseFinding/index.html
}
\examples{
data(dat3m)
## maybe str(dat3m) ; plot(dat3m) ...
}
\keyword{datasets}
