\name{r2.funct}
\alias{r2.funct}

\title{R2 function}

\description{ Calculates the R2 value }

\usage{ r2.funct(y, y.new) }

\arguments{
  \item{y}{ observed y values}
  \item{y.new}{ predicted y values}
}

\details{No details required }

\value{
A numeric value for the coefficient of determination R2 }

\references{
Dobson An introduction to linear modelling
}

\author{ Georgia Tsiliki }

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

r2.funct(1:10,1:10)
}

\keyword{ r2 }
