\name{suggest.trials.noxy}
\alias{suggest.trials.noxy}

\title{Returns suggested trials for a factorial design}

\description{Suggested trials are returned as a list with one object, a numeric matrix with the last column indicating with 1 the suggested trials. A column with NA values is added for the value sof the new experiments to be added.}

\usage{
suggest.trials.noxy(dataset, rawModel, additionalInfo)
}

\arguments{
  \item{dataset}{ Data for prediction. A list of two objects: datasetURI (a character string ), dataEntry (a data frame).}
  \item{rawModel}{ R model serialized (suggested trials for the data matrix supplied in exp.design.funct1). 
}
  \item{additionalInfo}{ Any additional information needed for rawModel. Here the list generated by exp.design.funct.noxy. The list should contain a field named 'predictedFeatures' which should be exactly the same as that returned by exp.design.funct.noxy function. 
}
}

\details{ No further details required }

\value{A list of one objected called 'predictions' which is also a list of one cell data-frames each containing the suggested trials for the data tested. This object is a matrix for the suggested trials per variable and level; the last column indicates with 1 the suggested trial
}

\references{ No references required }

\author{Georgia Tsiliki}

\note{ No notes for this function }


\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.


data("dat2m")
data("dat2i")

pred.res<- suggest.trials.noxy(null, dat2m, dat2i) 

}

\keyword{ suggestTFF }
