\name{suggest.trials.xy}
\alias{suggest.trials.xy}

\title{Returns suggested trials when data are available}

\description{Suggested trials are returned as a list with one object, a binary vector where 1 indicates suggested trial}

\usage{
suggest.trials.xy(dataset, rawModel, additionalInfo)
}

\arguments{
  \item{dataset}{ Data for prediction. A list of two objects: datasetURI (a character string ), dataEntry (a data frame).}
  \item{rawModel}{ R model serialized (suggested trials for the data matrix supplied in exp.design.funct1). 
}
  \item{additionalInfo}{ Any additional information needed for rawModel. Here the list generated by exp.design.funct.xy. The list should contain a field named 'predictedFeatures' which should be exactly the same as that returned by exp.design.funct.xy function. 
}
}

\details{ No further details required }

\value{A list of one objected called 'predictions' which is also a list of one cell data-frames each containing the suggested trials for the data tested. This object is a binary vector with 1 indicates suggested trial
}

\references{ No references required }

\author{Georgia Tsiliki}

\note{ No notes for this function }

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--  or do  help(data=index)  for the standard data sets.

data("dat1p")
data("dat1m")
data("dat1i")

pred.res<- suggest.trials.xy(dat1p, dat1m, dat1i) 

}

\keyword{ suggestT }
